Removed notExists chaining for updating shell script

+++ data/update_clustering.sh
@@ -1,4 +1,5 @@
 #!/bin/bash -e
+
 function fail() {
     echo "Error: $1"
     exit 1
@@ -57,10 +58,11 @@ NEWDB="$(abspath $2)" #"../data/targetDB"
 TMP="$(abspath $5)" #"tmp/"
 NEWCLUST="$(abspath $4)"
 
-notExists "$TMP/removedSeqs" \
-    && $MMSEQS diffseqdbs "$OLDDB" "$NEWDB" \
+if notExists "$TMP/removedSeqs"; then
+    $MMSEQS diffseqdbs "$OLDDB" "$NEWDB" \
                           "$TMP/removedSeqs" "$TMP/mappingSeqs" "$TMP/newSeqs" ${DIFF_PAR} \
-    || fail "Diff died"
+        || fail "Diff died"
+fi
 
 if [ ! -s "$TMP/mappingSeqs" ]; then
     echo <<WARN
@@ -80,28 +82,34 @@ if [ -n "$PRESERVE_REPR" ] && [ -f "$TMP/removedSeqs" ]; then
     echo "========= Recover removed representatives ========="
     echo "==================================================="
 
-    notExists "$TMP/OLDCLUST.allRepr" && ( \
+    if notExists "$TMP/OLDCLUST.allRepr"; then
+        ( \
             $MMSEQS result2stats "$OLDDB" "$OLDDB" "$OLDCLUST" "$TMP/OLDCLUST.allRepr" --stat firstline; \
             rm "$TMP/OLDCLUST.allRepr.index"; \
             tr -d '\000' < "$TMP/OLDCLUST.allRepr" > "$TMP/OLDCLUST.allReprNN"; \
             mv -f "$TMP/OLDCLUST.allReprNN" "$TMP/OLDCLUST.allRepr" \
         ) || fail "result2stats died"
+    fi
 
     LC_ALL=C comm -12 <(LC_ALL=C sort -T "$TMP" "$TMP/removedSeqs") <(LC_ALL=C sort -T "$TMP" "$TMP/OLDCLUST.allRepr") > "$TMP/OLDCLUST.removedRepr"
     if [[ -f "$TMP/OLDCLUST.removedRepr" ]]; then
-        notExists "$TMP/OLDCLUST.removedReprSeqs" \
-            && $MMSEQS createsubdb "$TMP/OLDCLUST.removedRepr" "$OLDDB" "$TMP/OLDDB.removedReprSeqs" \
-            || fail "createsubdb died"
+        if notExists "$TMP/OLDCLUST.removedReprSeqs"; then
+            $MMSEQS createsubdb "$TMP/OLDCLUST.removedRepr" "$OLDDB" "$TMP/OLDDB.removedReprSeqs" \
+                || fail "createsubdb died"
+        fi
 
-        notExists "$TMP/OLDCLUST.removedReprMapping" && ( \
+        if notExists "$TMP/OLDCLUST.removedReprMapping"; then
+            ( \
                 HIGHESTID="$(sort -T "$TMP" -r -n -k1,1 "${NEWDB}.index"| head -n 1 | cut -f1)"; \
                 awk -v highest="$HIGHESTID" \
                     'BEGIN { start=highest+1 } { print $1"\t"highest; highest=highest+1; }' \
                     "$TMP/OLDCLUST.removedRepr" > "$TMP/OLDCLUST.removedReprMapping"; \
                 cat "$TMP/OLDCLUST.removedReprMapping" >> "$TMP/mappingSeqs"; \
             ) || fail "Could not create $TMP/OLDCLUST.removedReprMapping"
+        fi
 
-        notExists "$TMP/NEWDB.withOldRepr" && ( \
+        if notExists "$TMP/NEWDB.withOldRepr"; then
+            ( \
                 ln -sf "$OLDDB" "$TMP/OLDDB.removedReprDb"; \
                 joinAndReplace "${OLDDB}.index" "$TMP/OLDDB.removedReprDb.index" "$TMP/OLDCLUST.removedReprMapping" "1.2 2.2 2.3"; \
                 joinAndReplace "${OLDDB}.lookup" "$TMP/OLDDB.removedReprDb.lookup" "$TMP/OLDCLUST.removedReprMapping" "1.2 2.2"; \
@@ -111,6 +119,7 @@ if [ -n "$PRESERVE_REPR" ] && [ -f "$TMP/removedSeqs" ]; then
                 cat "${NEWDB}.lookup" "$TMP/OLDDB.removedReprDb.lookup" > "$TMP/NEWDB.withOldRepr.lookup"; \
                 NEWDB="$TMP/NEWDB.withOldRepr"; \
             ) || fail "Could not create $TMP/NEWDB.withOldRepr"
+        fi
 
         if [ -n "$REMOVE_TMP" ]; then
             echo "Remove temporary files 1/3"
@@ -125,7 +134,8 @@ echo "==================================================="
 echo "=== Update the new sequences with the old keys ===="
 echo "==================================================="
 
-notExists "$TMP/newMappingSeqs" && ( \
+if notExists "$TMP/newMappingSeqs"; then
+    ( \
         OLDHIGHESTID="$(sort -T "$TMP" -r -n -k1,1 "${OLDDB}.index"| head -n 1 | cut -f1)"; \
         NEWHIGHESTID="$(sort -T "$TMP" -r -n -k1,1 "${NEWDB}.index"| head -n 1 | cut -f1)"; \
         MAXID="$(($OLDHIGHESTID>$NEWHIGHESTID?$OLDHIGHESTID:$NEWHIGHESTID))"; \
@@ -136,21 +146,23 @@ notExists "$TMP/newMappingSeqs" && ( \
         cat "$TMP/mappingSeqs.reverse" "$TMP/newSeqs.mapped" > "$TMP/newMappingSeqs"; \
         awk '{ print $2 }' "$TMP/newSeqs.mapped" > "$TMP/newSeqs"; \
     ) || fail "Could not create $TMP/newMappingSeqs"
+fi
+
+if notExists "$TMP/NEWDB.index"; then
+    joinAndReplace "${NEWDB}.index" "$TMP/NEWDB.index" "$TMP/newMappingSeqs" "1.2 2.2 2.3" \
+        || fail "join died"
+fi
 
 
-notExists "$TMP/NEWDB.index" \
-    && joinAndReplace "${NEWDB}.index" "$TMP/NEWDB.index" "$TMP/newMappingSeqs" "1.2 2.2 2.3" \
-    || fail "join died"
+if notExists "$TMP/NEWDB_h.index"; then
+    joinAndReplace "${NEWDB}_h.index" "$TMP/NEWDB_h.index" "$TMP/newMappingSeqs" "1.2 2.2 2.3" \
+        || fail "join died"
+fi
 
-
-notExists "$TMP/NEWDB_h.index" \
-    && joinAndReplace "${NEWDB}_h.index" "$TMP/NEWDB_h.index" "$TMP/newMappingSeqs" "1.2 2.2 2.3" \
-    || fail "join died"
-
-
-notExists "$TMP/NEWDB.lookup" \
-    && joinAndReplace "${NEWDB}.lookup" "$TMP/NEWDB.lookup" "$TMP/newMappingSeqs" "1.2 2.2" \
-    || fail "join died"
+if notExists "$TMP/NEWDB.lookup"; then
+    joinAndReplace "${NEWDB}.lookup" "$TMP/NEWDB.lookup" "$TMP/newMappingSeqs" "1.2 2.2" \
+        || fail "join died"
+fi
 
 ln -s "${NEWDB}" "$TMP/NEWDB"
 ln -s "${NEWDB}_h" "$TMP/NEWDB_h"
@@ -166,17 +178,18 @@ fi
 echo "==================================================="
 echo "====== Filter out the new from old sequences ======"
 echo "==================================================="
-notExists "$TMP/NEWDB.newSeqs" \
-    && $MMSEQS createsubdb "$TMP/newSeqs" "$NEWDB" "$TMP/NEWDB.newSeqs" \
-    || fail "Order died"
-
+if notExists "$TMP/NEWDB.newSeqs"; then
+    $MMSEQS createsubdb "$TMP/newSeqs" "$NEWDB" "$TMP/NEWDB.newSeqs" \
+        || fail "Order died"
+fi
 #read -n1
 echo "==================================================="
 echo "======= Extract representative sequences =========="
 echo "==================================================="
-notExists "$TMP/OLDDB.mapped.repSeq" \
-    && $MMSEQS result2msa "$OLDDB" "$OLDDB" "$OLDCLUST" "$TMP/OLDDB.repSeq" --only-rep-seq \
-    || fail "Result2msa died"
+if notExists "$TMP/OLDDB.mapped.repSeq"; then
+    $MMSEQS result2msa "$OLDDB" "$OLDDB" "$OLDCLUST" "$TMP/OLDDB.repSeq" --only-rep-seq \
+        || fail "Result2msa died"
+fi
 
 #read -n1
 echo "==================================================="
@@ -184,74 +197,83 @@ echo "======== Search the new sequences against ========="
 echo "========= previous (rep seq of) clusters =========="
 echo "==================================================="
 mkdir -p "$TMP/search"
-notExists "$TMP/newSeqsHits" \
-    && $RUNNER $MMSEQS search "$TMP/NEWDB.newSeqs" "$TMP/OLDDB.repSeq" "$TMP/newSeqsHits" "$TMP/search" --max-seqs 1 ${SEARCH_PAR} \
-    || fail "Search died"
+if notExists "$TMP/newSeqsHits"; then
+    $RUNNER $MMSEQS search "$TMP/NEWDB.newSeqs" "$TMP/OLDDB.repSeq" "$TMP/newSeqsHits" "$TMP/search" --max-seqs 1 ${SEARCH_PAR} \
+        || fail "Search died"
+fi
 
-notExists "$TMP/newSeqsHits.swapped.all" \
-    && $MMSEQS swapresults "$TMP/NEWDB.newSeqs" "$TMP/OLDDB.repSeq" "$TMP/newSeqsHits" "$TMP/newSeqsHits.swapped.all" \
-    || fail "Swapresults died"
+if notExists "$TMP/newSeqsHits.swapped.all"; then
+    $MMSEQS swapresults "$TMP/NEWDB.newSeqs" "$TMP/OLDDB.repSeq" "$TMP/newSeqsHits" "$TMP/newSeqsHits.swapped.all" \
+        || fail "Swapresults died"
+fi
 
-notExists "$TMP/newSeqsHits.swapped" \
-    && $MMSEQS filterdb "$TMP/newSeqsHits.swapped.all" "$TMP/newSeqsHits.swapped" --trim-to-one-column \
-    || fail "Trimming died"
+if notExists "$TMP/newSeqsHits.swapped"; then
+    $MMSEQS filterdb "$TMP/newSeqsHits.swapped.all" "$TMP/newSeqsHits.swapped" --trim-to-one-column \
+        || fail "Trimming died"
+fi
 
 #read -n1
 echo "==================================================="
 echo "=  Merge found sequences with previous clustering ="
 echo "==================================================="
 if [ -f "$TMP/newSeqsHits.swapped" ]; then
-    notExists "$TMP/updatedClust" \
-        && $MMSEQS mergedbs "$OLDCLUST" "$TMP/updatedClust" "$TMP/newSeqsHits.swapped" "$OLDCLUST" \
-        || fail "Mergeffindex died"
+    if notExists "$TMP/updatedClust"; then
+        $MMSEQS mergedbs "$OLDCLUST" "$TMP/updatedClust" "$TMP/newSeqsHits.swapped" "$OLDCLUST" \
+            || fail "Mergeffindex died"
+    fi
 else
-    notExists "$TMP/updatedClust" \
-        && ln -s "$OLDCLUST" "$TMP/updatedClust" \
-        || fail "Mv Oldclust to update died"
-
-    notExists "$TMP/updatedClust.index" \
-        && ln -s "$OLDCLUST.index" "$TMP/updatedClust.index" \
-        || fail "Mv Oldclust to update died"
+    if notExists "$TMP/updatedClust"; then
+        ln -s "$OLDCLUST" "$TMP/updatedClust" \
+            || fail "Mv Oldclust to update died"
+    fi
+    if notExists "$TMP/updatedClust.index"; then
+        ln -s "$OLDCLUST.index" "$TMP/updatedClust.index" \
+            || fail "Mv Oldclust to update died"
+    fi
 fi
 
 #read -n1
 echo "==================================================="
 echo "=========== Extract unmapped sequences ============"
 echo "==================================================="
-notExists "$TMP/noHitSeqList" \
-    && awk '$3==1 {print $1}' "$TMP/newSeqsHits.index" > "$TMP/noHitSeqList"
-
-notExists "$TMP/toBeClusteredSeparately" \
-    && $MMSEQS createsubdb "$TMP/noHitSeqList" "$NEWDB" "$TMP/toBeClusteredSeparately" \
-    || fail "Order of no hit seq. died"
-
+if notExists "$TMP/noHitSeqList"; then
+    awk '$3==1 {print $1}' "$TMP/newSeqsHits.index" > "$TMP/noHitSeqList" \
+        || fail "awk died"
+fi
+if notExists "$TMP/toBeClusteredSeparately"; then
+    $MMSEQS createsubdb "$TMP/noHitSeqList" "$NEWDB" "$TMP/toBeClusteredSeparately" \
+        || fail "Order of no hit seq. died"
+fi
 #read -n1
 echo "==================================================="
 echo "===== Cluster separately the alone sequences ======"
 echo "==================================================="
 
 mkdir -p "$TMP/cluster"
-notExists "$TMP/newClusters" \
-    && $MMSEQS cluster "$TMP/toBeClusteredSeparately" "$TMP/newClusters" "$TMP/cluster" ${CLUST_PAR} \
-    || fail "Clustering of new seq. died"
-
+if notExists "$TMP/newClusters"; then
+    $MMSEQS cluster "$TMP/toBeClusteredSeparately" "$TMP/newClusters" "$TMP/cluster" ${CLUST_PAR} \
+        || fail "Clustering of new seq. died"
+fi
 #read -n1
 echo "==================================================="
 echo "==== Merge the updated clustering together with ==="
 echo "=====         the new clusters               ======"
 echo "==================================================="
 if [ -f "$TMP/newClusters" ]; then
-    notExists "$NEWCLUST" \
-        && $MMSEQS concatdbs "$TMP/updatedClust" "$TMP/newClusters" "$NEWCLUST" \
-        || fail "Dbconcat died"
+    if notExists "$NEWCLUST"; then
+        $MMSEQS concatdbs "$TMP/updatedClust" "$TMP/newClusters" "$NEWCLUST" \
+            || fail "Dbconcat died"
+    fi
 else
-    notExists "$NEWCLUST" \
-        && mv "$TMP/updatedClust" "$NEWCLUST" \
-        || fail "Mv died"
+    if notExists "$NEWCLUST"; then
+        mv "$TMP/updatedClust" "$NEWCLUST" \
+            || fail "Mv died"
+    fi
 
-    notExists "${NEWCLUST}.index" \
-        && mv "$TMP/updatedClust.index" "${NEWCLUST}.index" \
-        || fail "Mv died"
+    if notExists "${NEWCLUST}.index"; then
+        mv "$TMP/updatedClust.index" "${NEWCLUST}.index" \
+            || fail "Mv died"
+    fi
 fi
 
 #read -n1
